# backend/src/main/resources/openapi-docs.yaml
openapi: 3.0.1
info:
  title: API Ristorino
  description: |
    Portal gastronómico para descubrir y reservar en restaurantes de Córdoba.
    
    ## Autenticación
    
    La API usa autenticación JWT (JSON Web Tokens) stateless. Para acceder a endpoints protegidos:
    
    1. Registrar nuevo cliente con `POST /api/clientes/register` o hacer login con `POST /api/clientes/login`
    2. Obtener el token JWT de la respuesta
    3. Incluir el token en el header de las requests: `Authorization: Bearer <token>`
    
    ## Logout
    
    Con JWT stateless no hay sesión del lado del servidor. El logout se realiza del lado del cliente:
    - Eliminar el token JWT del almacenamiento local (localStorage, cookies, etc.)
    - No hay endpoint de logout en el servidor ya que el token es independiente del servidor
  version: 1.0.0
  contact:
    name: Equipo DAS - UBP
    email: soporte@ristorino.com
servers:
  - url: http://localhost:8080
    description: Servidor de Desarrollo
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido desde POST /api/clientes/register o POST /api/clientes/login
...
security:
  - BearerAuth: []
paths:
  /api/clientes/register:
    post:
      tags:
        - Clientes
      summary: Registrar nuevo cliente
      description: |
        Crea un nuevo cliente y retorna un token JWT automáticamente.
        
        El token JWT debe incluirse en el header `Authorization: Bearer <token>` 
        para acceder a endpoints protegidos.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - apellido
                - nombre
                - password
                - correo
                - nroLocalidad
              properties:
                apellido:
                  type: string
                  maxLength: 120
                nombre:
                  type: string
                  maxLength: 120
                password:
                  type: string
                  minLength: 6
                  maxLength: 100
                correo:
                  type: string
                  format: email
                  maxLength: 150
                telefonos:
                  type: string
                  maxLength: 120
                nroLocalidad:
                  type: string
                  maxLength: 36
      responses:
        '201':
          description: Cliente registrado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Token JWT para autenticación
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  nroCliente:
                    type: string
                    format: uuid
                    description: UUID del cliente
                  nombre:
                    type: string
                    description: Nombre del cliente
                  apellido:
                    type: string
                    description: Apellido del cliente
                  correo:
                    type: string
                    format: email
                    description: Correo electrónico del cliente
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                nroCliente: "12345678-1234-1234-1234-123456789012"
                nombre: "Juan"
                apellido: "Pérez"
                correo: "juan.perez@example.com"
        '400':
          description: Datos inválidos o correo duplicado
        '500':
          description: Error interno del servidor
  /api/clientes/login:
    post:
      tags:
        - Clientes
      summary: Iniciar sesión
      description: |
        Autentica un cliente y retorna un token JWT.
        
        El token JWT debe incluirse en el header `Authorization: Bearer <token>` 
        para acceder a endpoints protegidos.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - correo
                - password
              properties:
                correo:
                  type: string
                  format: email
                  example: "juan.perez@example.com"
                password:
                  type: string
                  example: "miPassword123"
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Token JWT para autenticación
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  nroCliente:
                    type: string
                    format: uuid
                    description: UUID del cliente
                  nombre:
                    type: string
                    description: Nombre del cliente
                  apellido:
                    type: string
                    description: Apellido del cliente
                  correo:
                    type: string
                    format: email
                    description: Correo electrónico del cliente
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                nroCliente: "12345678-1234-1234-1234-123456789012"
                nombre: "Juan"
                apellido: "Pérez"
                correo: "juan.perez@example.com"
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Credenciales inválidas"
        '500':
          description: Error interno del servidor
...
  /api/contenidos/generar:
    post:
      tags:
        - Contenidos IA
      summary: Generar contenido publicitario con IA e integrar con SOAP
      description: |
        🤖 **Funcionalidad Innovadora con Inteligencia Artificial e Integración SOAP**
        
        Este endpoint utiliza ChatGPT de OpenAI para generar automáticamente textos publicitarios 
        personalizados para restaurantes y los sincroniza con el servicio SOAP del restaurante.
        
        **Flujo completo:**
        1. 📊 Analiza la información del restaurante desde la base de datos (tipo de comida, 
           ubicación, ambiente, horarios, etc.)
        2. 🎯 Construye un prompt contextualizado con todos los detalles relevantes
        3. 🤖 Envía el prompt a ChatGPT usando el **Prompt ID configurado** en OpenAI
        4. 💾 Guarda el contenido generado en `das_ristorino.contenidos_restaurantes`
        5. 📡 Notifica al servicio SOAP (`das-restaurante-soap`) para guardar en su BD
        6. 🔗 Actualiza `cod_contenido_restaurante` con el UUID retornado por el SOAP
        7. 📅 Establece vigencia automática de 1 mes
        
        **Integración SOAP:**
        - El contenido se sincroniza automáticamente con `das-restaurante-soap`
        - El servicio SOAP genera su propio UUID para el contenido
        - El campo `cod_contenido_restaurante` en ristorino almacena el UUID del SOAP
        - Esto permite referenciar el contenido del SOAP para notificaciones de clicks
        
        **Características:**
        - 🎨 Tono adaptativo según el ambiente del restaurante
        - 📏 Máximo 300 palabras
        - 🌐 **Soporte multiidioma**: El texto se genera en el idioma especificado por `nroIdioma`
          - Consulta la tabla `idiomas` para obtener `cod_idioma` (ej: es-AR, en-US, pt-BR)
          - El prompt incluye instrucciones para generar el texto en el idioma correcto
          - Ejemplos: español de Argentina, inglés, portugués de Brasil, etc.
        - 🔗 Integración automática REST → SOAP
        - 📊 Integración con sistema de clicks y promociones
        - 🎯 **Prompt ID por defecto**: `pmpt_68f93394cf6c8195955e0767742b9d7f05a21f383241fa79`
        
        **Prompt ID:**
        - Si NO especificas `promptId` → usa el configurado por defecto en `application.properties`
        - Si especificas `promptId` → usa el prompt que indiques (sobreescribe el default)
        
        **Seguridad:**
        - Endpoint público para pruebas (no requiere autenticación)
      security: []
      operationId: generarContenido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nroRestaurante
                - nroIdioma
              properties:
                nroRestaurante:
                  type: string
                  format: uuid
                  maxLength: 36
                  description: UUID del restaurante (obligatorio)
                  example: "5EF73536-5316-4D1E-9AD1-E78BEA0F4273"
                nroSucursal:
                  type: string
                  format: uuid
                  maxLength: 36
                  description: UUID de la sucursal (opcional, si no se especifica se usa info general)
                  example: "A1B2C3D4-E5F6-7890-1234-567890ABCDEF"
                nroIdioma:
                  type: string
                  format: uuid
                  maxLength: 36
                  description: |
                    UUID del idioma (obligatorio).
                    El sistema consulta la tabla `idiomas` para obtener el código del idioma (cod_idioma, ej: es-AR, en-US, pt-BR)
                    y genera el contenido publicitario en ese idioma.
                    El texto generado por OpenAI será en el idioma correspondiente al `nroIdioma` especificado.
                  example: "F1E2D3C4-B5A6-9876-5432-10FEDCBA9876"
                contextoAdicional:
                  type: string
                  maxLength: 500
                  description: Información adicional para el prompt (opcional)
                  example: "Destacar la nueva carta de vinos premium y el ambiente romántico"
                promptId:
                  type: string
                  maxLength: 100
                  description: |
                    ID del prompt guardado en OpenAI (opcional).
                    Si no se especifica, usa el prompt por defecto configurado en application.properties:
                    `pmpt_68f93394cf6c8195955e0767742b9d7f05a21f383241fa79`
                  example: "pmpt_68f93394cf6c8195955e0767742b9d7f05a21f383241fa79"
            examples:
              restaurante_general:
                summary: Generar con prompt por defecto (mínimo)
                value:
                  nroRestaurante: "5EF73536-5316-4D1E-9AD1-E78BEA0F4273"
                  nroIdioma: "F1E2D3C4-B5A6-9876-5432-10FEDCBA9876"
              sucursal_especifica:
                summary: Generar para sucursal específica con contexto
                value:
                  nroRestaurante: "5EF73536-5316-4D1E-9AD1-E78BEA0F4273"
                  nroSucursal: "A1B2C3D4-E5F6-7890-1234-567890ABCDEF"
                  nroIdioma: "F1E2D3C4-B5A6-9876-5432-10FEDCBA9876"
                  contextoAdicional: "Mencionar la experiencia única y el nuevo menú degustación"
              con_prompt_personalizado:
                summary: Usar un prompt diferente al configurado
                value:
                  nroRestaurante: "5EF73536-5316-4D1E-9AD1-E78BEA0F4273"
                  nroIdioma: "F1E2D3C4-B5A6-9876-5432-10FEDCBA9876"
                  promptId: "pmpt_OTRO_PROMPT_ID_DIFERENTE"
                  contextoAdicional: "Enfocado en eventos corporativos y reuniones de negocios"
      responses:
        '201':
          description: Contenido generado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  nroRestaurante:
                    type: string
                    format: uuid
                    example: "5EF73536-5316-4D1E-9AD1-E78BEA0F4273"
                  nroSucursal:
                    type: string
                    format: uuid
                    example: "A1B2C3D4-E5F6-7890-1234-567890ABCDEF"
                  nroIdioma:
                    type: string
                    format: uuid
                    example: "F1E2D3C4-B5A6-9876-5432-10FEDCBA9876"
                  nroContenido:
                    type: string
                    format: uuid
                    description: UUID del contenido recién generado
                    example: "X1Y2Z3W4-V5U6-7890-ABCD-EF1234567890"
                  nombreRestaurante:
                    type: string
                    example: "Los Aroza SRL"
                  nombreSucursal:
                    type: string
                    example: "Los Aroza - Centro"
                  contenidoGenerado:
                    type: string
                    description: Texto publicitario generado por ChatGPT y sincronizado con el SOAP del restaurante
                    example: "En el corazón de Córdoba, sobre la emblemática Av. Colón, Los Aroza te invita a vivir una experiencia gastronómica única. Nuestra sucursal Centro combina la exquisita tradición de la cocina japonesa con un ambiente íntimo y romántico..."
                  fechaIniVigencia:
                    type: string
                    format: date
                    description: Fecha de inicio de vigencia (automática)
                    example: "2025-10-29"
                  fechaFinVigencia:
                    type: string
                    format: date
                    description: Fecha de fin de vigencia (1 mes después)
                    example: "2025-11-29"
              example:
                nroRestaurante: "5EF73536-5316-4D1E-9AD1-E78BEA0F4273"
                nroSucursal: "A1B2C3D4-E5F6-7890-1234-567890ABCDEF"
                nroIdioma: "F1E2D3C4-B5A6-9876-5432-10FEDCBA9876"
                nroContenido: "X1Y2Z3W4-V5U6-7890-ABCD-EF1234567890"
                nombreRestaurante: "Los Aroza SRL"
                nombreSucursal: "Los Aroza - Centro"
                contenidoGenerado: "En el corazón de Córdoba, sobre la emblemática Av. Colón, Los Aroza te invita a vivir una experiencia gastronómica única. Nuestra sucursal Centro combina la exquisita tradición de la cocina japonesa con un ambiente íntimo y romántico, perfecto para compartir momentos especiales. Bajo una iluminación cálida y acompañados por música suave, nuestros chefs especializados elaboran cada pieza de sushi con dedicación artesanal..."
                fechaIniVigencia: "2025-10-29"
                fechaFinVigencia: "2025-11-29"
        '400':
          description: Datos inválidos o restaurante no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                restaurante_no_encontrado:
                  value:
                    error: "Restaurante no encontrado con ID: 5EF73536-5316-4D1E-9AD1-E78BEA0F4273"
                datos_invalidos:
                  value:
                    error: "El número de restaurante es obligatorio"
        '500':
          description: Error interno del servidor (OpenAI, BD o SOAP)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                openai_no_configurado:
                  value:
                    error: "La API key de OpenAI no está configurada. Configure la variable de entorno OPENAI_API_KEY."
                error_generacion:
                  value:
                    error: "Error al generar contenido con OpenAI: Rate limit exceeded"
                error_base_datos:
                  value:
                    error: "Error al guardar el contenido generado en la base de datos"
                error_soap:
                  value:
                    error: "Error en comunicación SOAP: Connection refused"
...